Create Recursive_Graphics.java
Currently does not allowing for scaling down of large images. Large images simply are too big to be displayed. Note that in line 47, the boolean operator determines if rectangles or ovals are to be used when breaking the image down. Furthermore, line 191 is where it can be changed. Currently all of this runs on console and needs to be changed to an actual GUI.